autofill/select-input.html

<!DOCTYPE HTML>
<head>
<meta charset="utf-8" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/autofill_helpers.js"></script>
</head>
<body>
<form id="test_form">
  <div>
    <label for="street">Street:</label>
    <input
      name="shipping-street"
      id="street"
      type="text"
      autocomplete="shipping address-line1"
    />
  </div>
  <div>
    <label for="country">Country:</label>
    <select
      name="shipping-country"
      id="country"
      type="text"
      autocomplete="shipping country"
    >
    <option value="US">United States</option>
    <option value="CA">Canada</option>
    <option value="MX">Mexico</option>
    <option value="DE">Germany</option>
    </select>
  </div>
</form>
<script>
  promise_test(async t => {
    if (!window.test_driver) {
      return;
    }
    const form = create_form_for_data({"address": "1 Main St", "country": "CA"});
    await test_driver.save_to_auto_fill(form);

    const suggestions = await test_driver.get_autofill_suggestions(document.getElementById("test_form"), document.getElementById("street"));

    // Assert that also non-visible fields of the same form were correctly autofilled:
    assert_element_has_value("country", "CA");
  });
</script>
